Rebuild my previous project.
This is a website for home page display. It is restructure https://github.com/RuotongX/bakewell-rtx by React In that repository, I notice that Nextjs has different form in development mode and production mode. That is caused by CSS loading orders are different. It is not easy for developers. So I rebuild that project by pure React. While I restructure the code, the page is look the same with production mode for original project, which the CSS is incorrect, so I rearrange my styling code and make CSS style normal.